Many changes
* Fix typo `oc/empty_env` -> `oci/empty_env`.
* Add missing alias in `projects/filebrowser-quantum`.
* Replace `%%NAME%%` with `alias` instead of project name.
* Change schedule time from 21 days to 1 week.
* Add `empty_makejail` parameter.

+++ scripts/update.sh
@@ -90,8 +90,10 @@ main()
     cp -a -- "${BASEDIR}/../template/workflow.yaml" "${wrksrc}/.github/workflows/build.yaml" || exit $?
     sed -i '' -Ee "s#%%NAME%%#${escape_project}#g" "${wrksrc}/.github/workflows/build.yaml" || exit $?
 
-    cp -a -- "${BASEDIR}/../template/Makejail" "${wrksrc}/Makejail" || exit $?
-    sed -i '' -Ee "s#%%NAME%%#${escape_project}#g" "${wrksrc}/Makejail" || exit $?
+    if [ ! -f "${projectdir}/empty_makejail" ]; then
+        cp -a -- "${BASEDIR}/../template/Makejail" "${wrksrc}/Makejail" || exit $?
+        sed -i '' -Ee "s#%%NAME%%#${param_alias}#g" "${wrksrc}/Makejail" || exit $?
+    fi
 
     mkdir -p -- "${wrksrc}/.daemonless" || exit $?
     printf "%s\n" "${param_daemonless}" > "${wrksrc}/.daemonless/config.yaml" || exit $?
